Fixed bugs with current user, firefox sesssion \w cookies and properly validations

+++ app/javascript/components/time_table/accounting_periods/period.js
@@ -30,7 +30,7 @@ class Period extends React.Component {
   }
 
   onDelete () {
-    this.props.onDelete(this.props.period.id);
+    if (confirm()) this.props.onDelete(this.props.period.id);
   }
 
   render () {
